% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_ari.R
\name{cluster_metric_ari}
\alias{cluster_metric_ari}
\title{Adjusted rand index}
\usage{
cluster_metric_ari(sce, cluster_col, truth_col)
}
\arguments{
\item{sce}{the SingleCellExperiment object}

\item{cluster_col}{the column name in colData(sce) containing the predicted
clusters}

\item{truth_col}{the column name in colData(sce) containing the true clusters}
}
\value{
Adjusted Rand index metric for comparing true cluster values and
  predicted cluster values.
}
\description{
To use this function in a CellBench workflow, it should be wrapped inside an
anoymous function. For example \code{function(x) { cluster_metric_ari(x,
""cluster_id", "cell_type") }}
}
