% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{cluster_metric_nmi}
\alias{cluster_metric_nmi}
\title{Normalized mutual information}
\usage{
cluster_metric_nmi(sce, cluster_col, truth_col)
}
\arguments{
\item{sce}{the SingleCellExperiment object}

\item{cluster_col}{the column name in colData(sce) containing the predicted
clusters}

\item{truth_col}{the column name in colData(sce) containing the true clusters}
}
\value{
Normalized mutual information for comparing true cluster values and
  predicted cluster values.
}
\description{
To use this function in a CellBench workflow, it should be wrapped inside an
anonymous function. For example \code{function(x) { cluster_metric_nmi(x,
"cluster_id", "cell_type") }}
}
