% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3.R
\name{cluster_sc3}
\alias{cluster_sc3}
\title{Clustering using SC3}
\usage{
cluster_sc3(sce, col.sym = "Symbol")
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}

\item{col.sym}{the column name containing gene symbols}
}
\description{
Clustering using SC3
}
