% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Clustering using Seurat}
\usage{
cluster_seurat(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}
}
\description{
Clustering using Seurat
}
