% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp.R
\name{cluster_sharp}
\alias{cluster_sharp}
\title{Clustering using SHARP}
\usage{
cluster_sharp(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
SingleCellExperiment object
}
\description{
Clustering using SHARP
}
