% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscan.R
\name{cluster_tscan}
\alias{cluster_tscan}
\title{Clustering using TSCAN}
\usage{
cluster_tscan(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}
}
\value{
vector of cluster identities
}
\description{
Clustering using TSCAN
}
