% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_id.R
\name{cluster_race_id}
\alias{cluster_race_id}
\title{Clustering using RaceID}
\usage{
cluster_race_id(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}
}
\value{
vector of cluster identities
}
\description{
Clustering using RaceID
}
